Move TracingProvider interface to `processor` package as discussed
